# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoMarket API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/related:
    get:
      operationId: listRelatedProducts
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # CATEGORIES
...
  /payments/submit:
    post:
      operationId: submitPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSubmission"
      responses:
        "201":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    patch:
      operationId: verifyPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerification"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /products/{id}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # MESSAGES
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    Product:
      type: object
      required: [id, name, slug, description, price, categoryId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        previewImages:
          type: array
          items:
            type: string
        demoUrl:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        changelog:
          type: ["string", "null"]
        documentation:
          type: ["string", "null"]
        videoPreviewUrl:
          type: ["string", "null"]
        status:
          type: string
        isFeatured:
          type: boolean
        salesCount:
          type: integer
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, description, price, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string
        previewImages:
          type: array
          items:
            type: string
        demoUrl:
          type: string
        downloadUrl:
          type: string
        version:
          type: string
        documentation:
          type: string
        videoPreviewUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string
        previewImages:
          type: array
          items:
            type: string
        demoUrl:
          type: string
        downloadUrl:
          type: string
        version:
          type: string
        changelog:
          type: string
        documentation:
          type: string
        videoPreviewUrl:
          type: string
        isFeatured:
          type: boolean
        status:
          type: string
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # CATEGORIES
...
    Payment:
      type: object
      required: [id, orderId, amount, chain, txHash, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        amount:
          type: number
        chain:
          type: string
        txHash:
          type: string
        screenshotUrl:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    PaymentSubmission:
      type: object
      required: [orderId, chain, txHash, amount]
      properties:
        orderId:
          type: integer
        chain:
          type: string
        txHash:
          type: string
        amount:
          type: number
        screenshotUrl:
          type: string
    PaymentVerification:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # MESSAGES
...