# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoMarket API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  shortDescription?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  tags?: string[];
  previewImages?: string[];
  demoUrl?: string;
  downloadUrl?: string;
  version?: string;
  documentation?: string;
  videoPreviewUrl?: string;
}